ユーザー登録画面を適切に分割
